// functions/src/generate-npc/npc-gen-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export interface Characteristics {
    /**
     * Strength
     */
    STR: number;
    /**
     * Dexterity
     */
    DEX: number;
    /**
     * Endurance
     */
    END: number;
    /**
     * Intelligence
     */
    INT: number;
    /**
     * Education
     */
    EDU: number;
    /**
     * Social
     */
    SOC: number;
}

/**
 * How exceptional are the characteristics of the NPC
 */
export type CitizenCategory = 'below_average' | 'average' | 'above_average' | 'exceptional';

export interface Error {
    message: string;
}

export type Experience = 'recruit' | 'rookie' | 'intermediate' | 'regular' | 'veteran' | 'elite';

/**
 * Gender. If you don't care, just omit it or use unspecified
 */
export type Gender = 'female' | 'male' | 'unspecified';

export interface NPC {
    first_name: string;
    surname: string;
    role: Role;
    citizen_category: CitizenCategory;
    experience: Experience;
    skills: string[];
    characteristics: Characteristics;
}

export interface NPCRequest {
    citizen_category?: CitizenCategory;
    experience?: Experience;
    gender?: Gender;
    role: Role;
}

/**
 * Role of the NPC
 */
export type Role = 'pilot' | 'navigator' | 'engineer' | 'steward' | 'medic' | 'marine' | 'gunner' | 'scout' | 'technician' | 'leader' | 'diplomat' | 'entertainer' | 'trader' | 'thug';

export interface GenerateNpcData {
    body: NPCRequest;
}

export type GenerateNpcResponse = (NPC);

export type GenerateNpcError = (Error);